2023.02.02(spring)

+++ backend/src/main/java/com/ssafy/db/repository/AttendanceRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<Attendance,Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance,Integer> {
 
     List<Attendance> findAttendanceByUseridAndGroupid(User user, Group group);
 }
