+++ backend/src/main/java/com/ssafy/db/repository/Group_MeetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Group_MeetRepository extends JpaRepository<Group_Meet, Long> {
+public interface Group_MeetRepository extends JpaRepository<Group_Meet, Integer> {
     Group_Meet findGroup_MeetByMeetid(int meetId);
 }
