+++ backend/src/main/java/com/ssafy/db/repository/MeetRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MeetRepository extends JpaRepository<Meet,Long> {
+public interface MeetRepository extends JpaRepository<Meet,Integer> {
     Meet findMeetByTitle(String title);
     Meet findMeetByUserid(int userId);
